gst/vaapi: move HEVC encoder test classes to lib
